requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
openpyxl==3.1.2
pandas==2.1.4
jdatetime==4.1.1
python-dotenv==1.0.0
Werkzeug==3.0.1
openai==1.18.0
flask-limiter==3.5.0
flask-swagger-ui==4.11.1
requests==2.31.0
pyotp==2.9.0
qrcode==7.4.2
Pillow==10.1.0
cryptography==41.0.7
